+++ src/cljs/rems/actions/review.cljs
@@ -1,24 +1,26 @@
 (ns rems.actions.review
   (:require [re-frame.core :as rf]
-            [rems.actions.action :refer [action-button action-form-view action-comment button-wrapper command!]]
+            [rems.actions.action :refer [action-attachment action-button action-comment action-form-view button-wrapper command!]]
             [rems.text :refer [text]]))
-
-(rf/reg-event-fx
- ::open-form
- (fn [{:keys [db]} _]
-   {:db (assoc db ::comment "")}))
-
-(rf/reg-sub ::comment (fn [db _] (::comment db)))
-(rf/reg-event-db ::set-comment (fn [db [_ value]] (assoc db ::comment value)))
 
 (def ^:private action-form-id "review")
 
 (rf/reg-event-fx
+ ::open-form
+ (fn [{:keys [db]} _]
+   {:db (assoc db ::comment "")
+    :dispatch [:rems.actions.action/set-attachment-id action-form-id nil]}))
+
+(rf/reg-sub ::comment (fn [db _] (::comment db)))
+(rf/reg-event-db ::set-comment (fn [db [_ value]] (assoc db ::comment value)))
+
+(rf/reg-event-fx
  ::send-review
- (fn [_ [_ {:keys [application-id comment on-finished]}]]
+ (fn [_ [_ {:keys [application-id comment attachments on-finished]}]]
    (command! :application.command/review
              {:application-id application-id
-              :comment comment}
+              :comment comment
+              :attachments attachments}
              {:description [text :t.actions/review]
               :collapse action-form-id
               :on-finished on-finished})
@@ -30,22 +32,28 @@
                   :on-click #(rf/dispatch [::open-form])}])
 
 (defn review-view
-  [{:keys [comment on-set-comment on-send]}]
+  [{:keys [application-id comment on-set-comment on-send]}]
   [action-form-view action-form-id
    (text :t.actions/review)
    [[button-wrapper {:id "review-button"
                      :text (text :t.actions/review)
                      :class "btn-primary"
                      :on-click on-send}]]
-   [action-comment {:id action-form-id
-                    :label (text :t.form/add-comments-not-shown-to-applicant)
-                    :comment comment
-                    :on-comment on-set-comment}]])
+   [:<>
+    [action-comment {:id action-form-id
+                     :label (text :t.form/add-comments-not-shown-to-applicant)
+                     :comment comment
+                     :on-comment on-set-comment}]
+    [action-attachment {:key action-form-id
+                        :application-id application-id}]]])
 
 (defn review-form [application-id on-finished]
-  (let [comment @(rf/subscribe [::comment])]
-    [review-view {:comment comment
+  (let [comment @(rf/subscribe [::comment])
+        attachments @(rf/subscribe [:rems.actions.action/attachments action-form-id])]
+    [review-view {:application-id application-id
+                  :comment comment
                   :on-set-comment #(rf/dispatch [::set-comment %])
                   :on-send #(rf/dispatch [::send-review {:application-id application-id
                                                          :comment comment
+                                                         :attachments attachments
                                                          :on-finished on-finished}])}]))