+++ src/cljs/rems/fields.cljs
@@ -246,56 +246,59 @@
                 [:label.form-check-label {:for option-id}
                  (localized label)]])))]))
 
+(defn attachment-link [attachment]
+  (when attachment
+    [:div.field
+     [:a.btn.btn-outline-secondary.mr-2
+      {:href (str "/applications/attachment/" (:attachment/id attachment))
+       :target :_blank}
+      (:attachment/filename attachment) " " [file-download]]]))
+
+(defn upload-button [id on-upload]
+  (let [upload-id (str id "-input")]
+    [:div.upload-file.mr-2
+     [:input {:style {:display "none"}
+              :type "file"
+              :id upload-id
+              :name upload-id
+              :accept ".pdf, .doc, .docx, .ppt, .pptx, .txt, image/*"
+              :on-change (fn [event]
+                           (let [filecontent (aget (.. event -target -files) 0)
+                                 form-data (doto (js/FormData.)
+                                             (.append "file" filecontent))]
+                             (on-upload form-data)))}]
+     [:button.btn.btn-outline-secondary
+      {:id id
+       :type :button
+       :on-click (fn [e] (.click (.getElementById js/document upload-id)))}
+      (text :t.form/upload)]]))
+
 (defn attachment-field
-  [{:keys [validation on-change on-set-attachment on-remove-attachment success] :as opts}]
+  [{:keys [validation on-set-attachment on-remove-attachment success] :as opts}]
   (let [title (localized (:field/title opts))
         value (:field/value opts)
         filename (get-in opts [:field/attachment :attachment/filename])
         upload-field-id (str (field-name opts) "-input")
-        click-upload (fn [e] (when-not (:readonly opts) (.click (.getElementById js/document upload-field-id))))
-        link (fn [attachment-id filename]
-               (when-not (empty? attachment-id)
-                 [:div.field
-                  [:a.btn.btn-outline-secondary.mr-2
-                   {:href (str "/applications/attachment/" attachment-id)
-                    :target :_blank}
-                   filename " " [file-download]]]))
-        upload-field [:div.upload-file.mr-2
-                      [:input {:style {:display "none"}
-                               :type "file"
-                               :id upload-field-id
-                               :name upload-field-id
-                               :accept ".pdf, .doc, .docx, .ppt, .pptx, .txt, image/*"
-                               :class (when validation "is-invalid")
-                               :on-change (fn [event]
-                                            (let [filecontent (aget (.. event -target -files) 0)
-                                                  filename (.-name filecontent)
-                                                  form-data (doto (js/FormData.)
-                                                              (.append "file" filecontent))]
-                                              (on-change (str filename " (" (localize-time (time/now)) ")"))
-                                              (on-set-attachment form-data title)))}]
-                      [:button.btn.btn-outline-secondary
-                       {:id (field-name opts)
-                        :type :button
-                        :on-click click-upload}
-                       (text :t.form/upload)]]
+        upload-field [upload-button
+                      (field-name opts)
+                      (fn [data]
+                        (on-set-attachment data title))]
         remove-button [:button.btn.btn-outline-secondary.mr-2
                        {:type :button
                         :on-click (fn [event]
-                                    (on-change "")
                                     (on-remove-attachment))}
                        (text :t.form/attachment-remove)]]
     [field-wrapper (assoc opts
-                          :readonly-component (link value filename)
+                          :readonly-component (attachment-link (:field/attachment opts))
                           :diff-component [:div {:style {:display :flex}}
                                            [:div
                                             (text :t.form/previous-value) ": "
-                                            (link (:field/previous-value opts) (get-in opts [:field/previous-attachment :attachment/filename]))]
+                                            (attachment-link (:field/previous-attachment opts))]
                                            [:div
                                             (text :t.form/current-value) ": "
-                                            (link value filename)]])
+                                            (attachment-link (:field/attachment opts))]])
      [:div.flex-row.d-flex.align-items-center
-      (link value filename)
+      (attachment-link (:field/attachment opts))
       (if (empty? value)
         upload-field
         remove-button)